chore: added benchmarking and ran tests
