+++ src/RecipeShare.sln
@@ -6,6 +6,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RecipeShare.Library", "Reci
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RecipeShare.Tests", "..\tests\RecipeShare.Tests\RecipeShare.Tests.csproj", "{A3FF3704-0F0B-4B2E-B2F4-69B7B38A71A0}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RecipeShare.Benchmarks", "..\tests\RecipeShare.Benchmarks\RecipeShare.Benchmarks.csproj", "{38D2A3DD-78F9-4C03-8328-F9E86123FF4F}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -24,5 +26,9 @@ Global
 		{A3FF3704-0F0B-4B2E-B2F4-69B7B38A71A0}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A3FF3704-0F0B-4B2E-B2F4-69B7B38A71A0}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A3FF3704-0F0B-4B2E-B2F4-69B7B38A71A0}.Release|Any CPU.Build.0 = Release|Any CPU
+		{38D2A3DD-78F9-4C03-8328-F9E86123FF4F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{38D2A3DD-78F9-4C03-8328-F9E86123FF4F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{38D2A3DD-78F9-4C03-8328-F9E86123FF4F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{38D2A3DD-78F9-4C03-8328-F9E86123FF4F}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
