+++ tests/RecipeShare.Benchmarks/BenchmarkDotNet.Artifacts/results/RecipeShare.Benchmarks.BenchmarkTests.RecipeEndpointBenchmarks-report.csv
@@ -1,0 +1,5 @@
+Method,Job,AnalyzeLaunchVariance,EvaluateOverhead,MaxAbsoluteError,MaxRelativeError,MinInvokeCount,MinIterationTime,OutlierMode,Affinity,EnvironmentVariables,Jit,LargeAddressAware,Platform,PowerPlanMode,Runtime,AllowVeryLargeObjects,Concurrent,CpuGroups,Force,HeapAffinitizeMask,HeapCount,NoAffinitize,RetainVm,Server,Arguments,BuildConfiguration,Clock,EngineFactory,NuGetReferences,Toolchain,IsMutator,InvocationCount,IterationCount,IterationTime,LaunchCount,MaxIterationCount,MaxWarmupIterationCount,MemoryRandomization,MinIterationCount,MinWarmupIterationCount,RunStrategy,UnrollFactor,WarmupCount,Mean,Error,StdDev,Median,Gen0,Gen1,Gen2,Allocated
+'500 Sequential Requests',Job-RISBLR,False,Default,Default,Default,Default,Default,Default,0000000000,Empty,RyuJit,Default,Arm64,8c5e7fda-e8bf-4a96-9a85-a6e23a8c635c,.NET 8.0,False,True,False,True,Default,Default,False,False,False,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,1,Default,Default,Default,Default,Default,Default,16,2,"1,196.049 ms",19.1574 ms,16.9826 ms,"1,195.385 ms",13000.0000,1000.0000,0.0000,107422.84 KB
+'500 Requests in Batches of 50',Job-RISBLR,False,Default,Default,Default,Default,Default,Default,0000000000,Empty,RyuJit,Default,Arm64,8c5e7fda-e8bf-4a96-9a85-a6e23a8c635c,.NET 8.0,False,True,False,True,Default,Default,False,False,False,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,1,Default,Default,Default,Default,Default,Default,16,2,305.136 ms,12.5742 ms,35.8748 ms,292.528 ms,15000.0000,8000.0000,3000.0000,113754.14 KB
+'500 Requests All Concurrent',Job-RISBLR,False,Default,Default,Default,Default,Default,Default,0000000000,Empty,RyuJit,Default,Arm64,8c5e7fda-e8bf-4a96-9a85-a6e23a8c635c,.NET 8.0,False,True,False,True,Default,Default,False,False,False,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,1,Default,Default,Default,Default,Default,Default,16,2,322.173 ms,9.6325 ms,27.1686 ms,313.922 ms,14000.0000,4000.0000,1000.0000,108807.39 KB
+'Single Request Baseline',Job-RISBLR,False,Default,Default,Default,Default,Default,Default,0000000000,Empty,RyuJit,Default,Arm64,8c5e7fda-e8bf-4a96-9a85-a6e23a8c635c,.NET 8.0,False,True,False,True,Default,Default,False,False,False,Default,Default,Default,Default,Default,Default,Default,Default,Default,Default,1,Default,Default,Default,Default,Default,Default,16,2,2.425 ms,0.0458 ms,0.0871 ms,2.410 ms,23.4375,3.9063,0.0000,216.82 KB
