+++ tests/RecipeShare.Benchmarks/BenchmarkDotNet.Artifacts/results/RecipeShare.Benchmarks.BenchmarkTests.RecipeEndpointBenchmarks-report.html
@@ -1,0 +1,34 @@
+<!DOCTYPE html>
+<html lang='en'>
+<head>
+<meta charset='utf-8' />
+<title>RecipeShare.Benchmarks.BenchmarkTests.RecipeEndpointBenchmarks-20250722-234513</title>
+
+<style type="text/css">
+	table { border-collapse: collapse; display: block; width: 100%; overflow: auto; }
+	td, th { padding: 6px 13px; border: 1px solid #ddd; text-align: right; }
+	tr { background-color: #fff; border-top: 1px solid #ccc; }
+	tr:nth-child(even) { background: #f8f8f8; }
+</style>
+</head>
+<body>
+<pre><code>
+BenchmarkDotNet v0.15.2, macOS Sequoia 15.5 (24F74) [Darwin 24.5.0]
+Apple M4, 1 CPU, 10 logical and 10 physical cores
+.NET SDK 8.0.411
+  [Host]     : .NET 8.0.17 (8.0.1725.26602), Arm64 RyuJIT AdvSIMD
+  Job-RISBLR : .NET 8.0.17 (8.0.1725.26602), Arm64 RyuJIT AdvSIMD
+</code></pre>
+<pre><code>LaunchCount=1  WarmupCount=2  
+</code></pre>
+
+<table>
+<thead><tr><th>Method                   </th><th>Mean  </th><th>Error</th><th>StdDev</th><th>Median</th><th>Gen0</th><th>Gen1</th><th>Gen2</th><th>Allocated</th>
+</tr>
+</thead><tbody><tr><td>&#39;500 Sequential Requests&#39;</td><td>1,196.049 ms</td><td>19.1574 ms</td><td>16.9826 ms</td><td>1,195.385 ms</td><td>13000.0000</td><td>1000.0000</td><td>-</td><td>107422.84 KB</td>
+</tr><tr><td>&#39;500 Requests in Batches of 50&#39;</td><td>305.136 ms</td><td>12.5742 ms</td><td>35.8748 ms</td><td>292.528 ms</td><td>15000.0000</td><td>8000.0000</td><td>3000.0000</td><td>113754.14 KB</td>
+</tr><tr><td>&#39;500 Requests All Concurrent&#39;</td><td>322.173 ms</td><td>9.6325 ms</td><td>27.1686 ms</td><td>313.922 ms</td><td>14000.0000</td><td>4000.0000</td><td>1000.0000</td><td>108807.39 KB</td>
+</tr><tr><td>&#39;Single Request Baseline&#39;</td><td>2.425 ms</td><td>0.0458 ms</td><td>0.0871 ms</td><td>2.410 ms</td><td>23.4375</td><td>3.9063</td><td>-</td><td>216.82 KB</td>
+</tr></tbody></table>
+</body>
+</html>
