+++ tests/RecipeShare.Benchmarks/Factories/CustomWebAppFactory.cs
@@ -1,0 +1,15 @@
+using Microsoft.AspNetCore.Mvc.Testing;
+using Microsoft.Extensions.Hosting;
+
+namespace RecipeShare.Factories;
+
+public class CustomWebAppFactory : WebApplicationFactory<Program>
+{
+    protected override IHost CreateHost(IHostBuilder builder)
+    {
+ 
+        var host = builder.Build();
+        host.Start();
+        return host;
+    }
+}
