+++ tests/RecipeShare.Benchmarks/RecipeShare.Benchmarks.csproj
@@ -1,0 +1,17 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <OutputType>Exe</OutputType>
+        <TargetFramework>net8.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+    </PropertyGroup>
+
+    <ItemGroup>
+      <PackageReference Include="BenchmarkDotNet" Version="0.15.2" />
+      <PackageReference Include="Microsoft.AspNetCore.Mvc.Testing" Version="8.0.18" />
+      <PackageReference Include="System.Net.Http" Version="4.3.4" />
+        <ProjectReference Include="..\..\src\RecipeShare.Api\RecipeShare.Api.csproj" />
+    </ItemGroup>
+
+</Project>
